Added Crimson Doll and Primordial Pearl recipes
Interacting with the Thaumcraft Outer Lands in skyblock can be very
RNG-heavy, with only a small chance to spawn Crimson Cultists when
placing a Portal-in-a-Box, which are necessary for late-game
progression. The Crimson Doll fixes that.

Also, generating multiple Outer Lands mazes can be glitchy, and they
permit at most one Primordial Pearl per maze. Adding the ability to
sieve the Ancient Stone in the Outer Lands for additional Primordial
Pearls (or Powerless Pearls if Witching Gadgets is loaded) eliminates
the need for multiple portals.